requirements.txt
langchain
langchain-community
boto3
pypdf
python-dotenv
streamlit
faiss-cpu